+++ src/controllers/game.controller.ts
@@ -1,11 +1,14 @@
 import { Request, Response, NextFunction, Router } from 'express';
 import gameModel, { Game } from '../models/game.model';
-import validateFile from '../util/fileBuffer';
-import filetype from 'file-type';
+import validateBuffer from '../storage/buffer.file';
 import { extImage as ImageType } from '../util/utilities';
+import handlerfile from '../storage/handler.file';
 import { env } from '../config/config';
-import { promises as fs } from 'fs';
+import filetype from 'file-type';
+import filesys, { promises as fs } from 'fs';
 import path from 'path';
+import del from 'del';
+
 
 class GameController {
     router: Router;
@@ -18,10 +21,12 @@ class GameController {
     public async getGames(req: Request, res: Response): Promise<void> {
         try {
             const games = await gameModel.find();
-            res.json({
-                msg: "Bienvenid@s a GamersApp",
-                games
-            });
+            if (games) {
+                res.json({
+                    msg: "Bienvenid@s a GamersApp",
+                    games
+                });
+            }
         } catch (e) {
             console.log(e);
             res.status(500).send('Error obteniendo todos los juegos');
@@ -32,9 +37,11 @@ class GameController {
     public async getGame(req: Request, res: Response): Promise<void> {
         try {
             const game = await gameModel.findById(req.body.gid);
-            res.json({
-                game
-            });
+            if (game) {
+                res.json({
+                    game
+                });
+            }
         } catch (e) {
             console.log(e);
             res.status(500).send('Error obteniendo el juego');
@@ -45,9 +52,11 @@ class GameController {
     public async getGamesUid(req: Request, res: Response): Promise<void> {
         try {
             const games = await gameModel.find({ uid: { $regex: req.body.uid } });
-            res.json({
-                games
-            });
+            if (games) {
+                res.json({
+                    games
+                });
+            }
         } catch (e) {
             console.log(e);
             res.status(500).send('Error obteniendo juegos para este usuario');
@@ -57,21 +66,22 @@ class GameController {
     //Validar número mágico de la imagen y subirla al servidor
     public async validateFile(req: Request, res: Response, next: NextFunction) {
         try {
-            await validateFile(req, res, async () => {
-                const gid = req.body.gid;                
-                const bufferFile = await filetype.fromBuffer(req.file.buffer);
-                let contentType: String[] = [ImageType.jpeg, ImageType.png, ImageType.jpg];
-                let isEquals: Boolean = false;
-                contentType.forEach((type) => {
-                    if (type === bufferFile?.mime) {
-                        isEquals = true;
+            await validateBuffer(req, res, async () => {
+                const gid = req.body.gid;
+                const contentType = await filetype.fromBuffer(req.file.buffer);
+                let mimeTypes: String[] = [ImageType.jpeg, ImageType.png, ImageType.jpg];
+                let isMatch: Boolean = false;
+                mimeTypes.forEach((type) => {
+                    if (type === contentType?.mime) {
+                        isMatch = true;
                         return;
                     }
                 });
-
-                if (isEquals) {
+                if (isMatch) {
                     if (gid === undefined) {
-                        await fs.writeFile(path.join(process.cwd(), 'uploads', req.file.originalname), req.file.buffer);
+                        const dirPath = await handlerfile.currentPatch('uploads');
+                        const newPath = await handlerfile.newPatch(req.file.originalname);
+                        await handlerfile.fileUpload(dirPath, newPath, req.file.buffer);
                     }
                 } else {
                     return res.status(415).send('El formato del archivo no es valido');
@@ -87,33 +97,21 @@ class GameController {
     //Crear juego
     public async createGame(req: Request, res: Response, next: NextFunction): Promise<void> {
         try {
-            const {
-                gname,
-                gdescription,
-                ggender,
-                gconsole,
-                grequirements,
-                gauthor,
-                uid
-            } = req.body;
-
+            const { gname, gdescription, ggender, gconsole, grequirements,
+                gauthor, uid } = req.body;
             const gimage = `/${env.desUpload}/${req.file.originalname}`;
 
             const game: Game = new gameModel({
-                gname,
-                gdescription,
-                ggender,
-                gconsole,
-                grequirements,
-                gauthor,
-                gimage,
-                uid
+                gname, gdescription, ggender, gconsole, grequirements,
+                gauthor, gimage, uid
             });
 
             const newGame = await game.save();
-            res.json({
-                game: newGame
-            });
+            if (newGame) {
+                res.json({
+                    game: newGame
+                });
+            }
         } catch (e) {
             console.log(e);
             res.status(500).send('Error registrando el juego');
@@ -122,61 +120,58 @@ class GameController {
 
     //Eliminar juego por id
     public async deleteGame(req: Request, res: Response): Promise<void> {
-        try {                     
+        try {
             const currentGame = await gameModel.findById(req.body.gid);
-            await gameModel.findByIdAndDelete(req.body.gid);            
-            if(currentGame?.gimage !== undefined){           
-            await fs.unlink(path.join(process.cwd(), currentGame?.gimage));
-        }
-            res.json({
-                message: "Este juego fue eliminado con éxito",
-            });
+            const game = await gameModel.findByIdAndDelete(req.body.gid);
+            if (game) {
+                if (currentGame?.gimage !== undefined) {
+                    if (filesys.existsSync(path.join(process.cwd(), currentGame?.gimage))) {
+                        await fs.unlink(path.join(process.cwd(), currentGame?.gimage));
+                    } else {
+                        res.json({
+                            message: "El archivo del juego no existe",
+                        });
+                    }
+                }
+                res.json({
+                    message: "Este juego fue eliminado con éxito",
+                });
+            }
         } catch (e) {
             console.log(e);
             res.status(500).send('Error eliminando el juego');
         }
     }
+
     //Actualizar juego
     public async updateGame(req: Request, res: Response): Promise<void> {
         try {
             const _id = req.body.gid;
-            const {
-                gname,
-                gdescription,
-                ggender,
-                gconsole,
-                grequirements,
-                gauthor,
-                uid
-            } = req.body;
-
-            let urlImage = null;
-            const currentGame = await gameModel.findById(_id);
-            urlImage = `/${env.desUpload}/${req.file.originalname}`;
-
-            if (urlImage !== currentGame?.gimage 
-                && currentGame?.gimage !== undefined) {                
-                await fs.unlink(path.join(process.cwd(), currentGame?.gimage));
-                fs.writeFile(path.join(process.cwd(), 'uploads', req.file.originalname), req.file.buffer);
-            }
+            const newImage = `/${env.desUpload}/${req.file.originalname}`;
+            const { gname, gdescription, ggender, gconsole, grequirements,
+                gauthor, uid } = req.body;
 
             const newGame: Game = new gameModel({
-                _id,
-                gname,
-                gdescription,
-                ggender,
-                gconsole,
-                grequirements,
-                gauthor,
-                gimage: urlImage,
-                uid
+                _id, gname, gdescription, ggender, gconsole, grequirements,
+                gauthor, gimage: newImage, uid
             });
+            const currentGame = await gameModel.findById(_id);
+            const game = await gameModel.findByIdAndUpdate(_id, newGame, { new: true });
 
-            const game = await gameModel.findByIdAndUpdate(req.body.gid, newGame, { new: true });
-            res.json({
-                message: "Juego actualizado con éxito",
-                user: game
-            });
+            if (game) {
+                if (newImage !== currentGame?.gimage
+                    && currentGame?.gimage !== undefined) {
+                    const currentPath = await handlerfile.currentPatch(currentGame?.gimage);
+                    const newPath = await handlerfile.newPatch(req.file.originalname);
+                    await handlerfile.deleteFileUpload(currentPath, newPath, req.file.buffer);
+                }
+                res.json({
+                    message: "Juego actualizado con éxito",
+                    user: game
+                });
+            } else {
+                res.status(404).send('Juego no encontrado');
+            }
         } catch (e) {
             console.log(e);
             res.status(500).send('Error actualizando el juego');
@@ -187,7 +182,10 @@ class GameController {
     public async deleteGames(req: Request, res: Response): Promise<void> {
         try {
             await gameModel.deleteMany({});
-            await fs.rmdir(path.join(process.cwd(), 'uploads'),{ recursive: true });
+            const urlPath = path.join(process.cwd(), 'uploads');
+            if (filesys.existsSync(urlPath)) {
+                await del(urlPath);
+            }
             res.json({
                 message: "Juegos eliminados con éxito",
             });
